+++ src/components/MapBox.vue
@@ -24,16 +24,7 @@
       </div>
     </div>
     <div id="mapContainer">
-      <div id="subtitle">
-        <div id="subtitleText">
-          <p>An Indicator of Natural Water Storage</p>
-        </div>
-        <div id="subtitleInforButton">
-          <a href="#" class="icon">
-            <font-awesome-icon icon="info" />
-          </a>
-        </div>
-      </div>
+      <MapSubtitle />
       <MapLegend :legend-title="legendTitle" />
       <MglMap
         id="map"
@@ -72,7 +63,8 @@
   </div>
 </template>
 <script>
-    import MapLegend from './MapLegend'
+    import MapSubtitle from './MapSubtitle';
+    import MapLegend from './MapLegend';
     import {
         MglMap,
         MglNavigationControl,
@@ -87,6 +79,7 @@
         name: 'MapBox',
         components: {
             MglMap,
+            MapSubtitle,
             MglNavigationControl,
             MglGeolocateControl,
             MglFullscreenControl,
@@ -244,43 +237,6 @@
   #mapContainer{
     position: relative;
     height:80vh;
-  }
-
-  #subtitle{
-    background:rgb(255,255,255);
-    background:rgba(255,255,255,.6);
-    position: absolute;
-    top: 10px;
-    left: 50%;
-    transform: translateX(-50%);
-    z-index: 1;
-    padding: 5px 10px;
-    border-radius: 5px;
-    display: flex;
-    flex-direction: row;
-    text-align: center;
-
-    p{
-      margin: 0;
-    }
-
-    #subtitleText{
-      flex: 1;
-      padding: 0 5px 0 0;
-      border-right: 1px solid #000;
-    }
-
-    #subtitleInforButton{
-      margin: 0 0 0 5px;
-      width: 20px;
-      
-
-      a{
-        color: #000;
-        display: block;
-        outline: none;
-      }
-    }
   }
 
   @media screen and (min-width:600px){